Move EditModal component into separate file
